refactor/pooling: add functions for pooling operations
