+++ src/frameworks/native/helper.rs
@@ -396,64 +396,3 @@ macro_rules! impl_ops_lrn_for {
         }
     );
 }
-
-#[macro_export]
-macro_rules! impl_ops_pooling_for {
-    ($t:ident, $b:ty) => (
-        impl ::plugin::Pooling<$t> for $b {
-            fn new_pooling_config(
-                &self,
-                window: &[i32],
-                padding: &[i32],
-                stride: &[i32]
-            ) -> Result<Self::CPOOL, ::co::error::Error> {
-                unimplemented!();
-                Ok(::frameworks::native::helper::PoolingConfig)
-            }
-
-            fn pooling_max(
-                &self,
-                x: &mut SharedTensor<$t>,
-                result: &mut SharedTensor<$t>,
-                config: &Self::CPOOL
-            ) -> Result<(), ::co::error::Error> {
-                unimplemented!();
-                Ok(())
-            }
-
-            fn pooling_max_plain(
-                &self,
-                x: &SharedTensor<$t>,
-                result: &mut SharedTensor<$t>,
-                config: &Self::CPOOL
-            ) -> Result<(), ::co::error::Error> {
-                unimplemented!();
-                Ok(())
-            }
-            #[allow(unused_variables)]
-            fn pooling_max_grad(
-                &self,
-                x: &mut SharedTensor<$t>,
-                x_diff: &mut SharedTensor<$t>,
-                result: &mut SharedTensor<$t>,
-                result_diff: &mut SharedTensor<$t>,
-                config: &Self::CPOOL
-            ) -> Result<(), ::co::error::Error> {
-                unimplemented!();
-                Ok(())
-            }
-
-            fn pooling_max_grad_plain(
-                &self,
-                x: &SharedTensor<$t>,
-                x_diff: &SharedTensor<$t>,
-                result: &SharedTensor<$t>,
-                result_diff: &mut SharedTensor<$t>,
-                config: &Self::CPOOL
-            ) -> Result<(), ::co::error::Error> {
-                unimplemented!();
-                Ok(())
-            }
-        }
-    );
-}
