+++ src/tests/pooling.rs
@@ -92,3 +92,11 @@ mod cuda {
     test_cuda!(test_pooling_max_grad, pooling_max_grad_f32, pooling_max_grad_f64);
 }
 
+mod native {
+    use super::*;
+    //test_native!(test_pooling_avg, pooling_avg_f32, pooling_avg_f64);
+    //test_native!(test_pooling_avg_grad, pooling_avg_grad_f32, pooling_avg_grad_f64);
+    test_native!(test_pooling_max, pooling_max_f32, pooling_max_f64);
+    test_native!(test_pooling_max_grad, pooling_max_grad_f32, pooling_max_grad_f64);
+}
+
